add: namespace for config setters to avoid littering the config class
