+++ Makefile
@@ -28,7 +28,8 @@ SRCS   	= \
         $(SRCS_DIR)/ConfigParser.cpp \
         $(SRCS_DIR)/ConfigDispatcher.cpp \
         $(SRCS_DIR)/Utils.cpp \
-		$(SRCS_DIR)/main_config.cpp
+        $(SRCS_DIR)/ConfigSetters.cpp \
+		$(SRCS_DIR)/main_config.cpp 
 TESTS	= \
