+++ src/Config.cpp
@@ -3,6 +3,7 @@
 #include "ConfigParser.hpp"
 #include "LocationInfo.hpp"
 #include "Utils.hpp"
+#include "ConfigSetters.hpp"
 #include <algorithm>
 #include <cstddef>
 #include <exception>
@@ -398,93 +399,16 @@ std::string	Config::get_error_page(const int key)
 	}
 }
 
-void	set_root(const std::vector <std::string>& root, LocationInfo*& new_location)
-{
-	if (root.empty() == false)
-	{
-		new_location->set_root(root[0]);
-	}
-}
-
-void	set_directory_listing(const std::vector <std::string>& directory_listing, LocationInfo*& new_location)
-{
-	if (directory_listing.empty() == false)
-	{
-		if (directory_listing[0] == "enabled")
-		{
-			new_location->set_directory_listing(true);
-		}
-	}
-}
-
-void	set_allowed_methods(const std::vector <std::string>& allowed_methods, LocationInfo*& new_location)
-{
-	if (allowed_methods.empty() == false)
-	{
-		new_location->set_allowed_methods(allowed_methods);
-	}
-	else 
-	{
-		std::vector <std::string> foo;
-		foo.push_back("none");
-		new_location->set_allowed_methods(allowed_methods);
-	}
-}
-
-void	set_return(const std::vector <std::string>& rtrn, LocationInfo*& new_location)
-{
-	if (rtrn.empty() == false)
-	{
-		new_location->set_return(rtrn[0]);
-	}
-}
-
-void	set_alias(const std::vector <std::string>& alias, LocationInfo*& new_location)
-{
-	if (alias.empty() == false)
-	{
-		new_location->set_alias(alias);
-	}
-}
-
-void	set_cgi_path(const std::vector <std::string>& cgi_path, LocationInfo*& new_location)
-{
-	if (cgi_path.empty() == false)
-	{
-		new_location->set_cgi_path(cgi_path[0]);
-	}
-}
-
-
-void	set_cgi_extension(const std::vector <std::string>& extension, LocationInfo*& new_location)
-{
-	if (extension.empty() == false)
-	{
-		new_location->set_cgi_extensions(extension);
-	}
-}
-
-void	set_autoindex(const std::vector <std::string>& autoindex, LocationInfo*& new_location)
-{
-	if (autoindex.empty() == false)
-	{
-		if (autoindex[0] == "enabled")
-		{
-			new_location->set_autoindex(true);
-		}
-	}
-}
-
 void	Config::initialize_location_setters()
 {
-	_location_setters["root"] = &set_root;
-	_location_setters["directory_listing"] = set_directory_listing;
-	_location_setters["allowed_methods"] = set_allowed_methods;
-	_location_setters["return"] = set_return;
-	_location_setters["alias"] = set_alias;
-	_location_setters["handler"] = set_cgi_path;
-	_location_setters["extension"] = set_cgi_extension;
-	_location_setters["autoindex"] = set_autoindex;
+	_location_setters["root"] = &Setters::set_root;
+	_location_setters["directory_listing"] = &Setters::set_directory_listing;
+	_location_setters["allowed_methods"] = &Setters::set_allowed_methods;
+	_location_setters["return"] = &Setters::set_return;
+	_location_setters["alias"] = &Setters::set_alias;
+	_location_setters["handler"] = &Setters::set_cgi_path;
+	_location_setters["extension"] = &Setters::set_cgi_extension;
+	_location_setters["autoindex"] = &Setters::set_autoindex;
 }
 
 Config::~Config()
